+++ ChangeLog.md
@@ -4,6 +4,7 @@
     - Properly updating the private message unread count
     - Improved channel conversation mode detection support
     - Added support for HTML encoded links (Example would be when copying from Edge the URL)
+    - Enabled context menus for all clickable client tags
     
 * **08.12.20**
     - Fixed the permission editor not resolving unique ids